Catch cuda driver shutdown error in NCCLWatchdog
There is a design flaw in NCCLWatchdog, namely it spawns threads that
talk to the CUDA api, but the CUDA api may have been deinitialized,
forming a race.

This is a known issue with widespread impact
(https://github.com/pytorch/pytorch/issues/90848).

A partial fix was landed already, but it applied too narrowly:
https://github.com/pytorch/pytorch/commit/ec071a0815adfd180ba1ab103be2f7d2227f07cc

This PR is a copy-paste of the previous fix, applying to one more case,
plugging a hole.  We probably need to do a more thorough review and
either plug all the holes, or design this differently.

ghstack-source-id: 81033736297edc64c93dd122296ab4bd11317796
Pull Request resolved: https://github.com/pytorch/pytorch/pull/106503